fix overflow I could just make it use u128 but considering that nobody should be rolling that high to begin with theres straight up no point.

+++ src/commands/utility/roll.rs
@@ -11,8 +11,8 @@ use rand::{prelude::Distribution, thread_rng};
 )]
 pub async fn roll(
     ctx: Context<'_>,
-    #[description = "Die size"] die_size: Option<u64>,
-    #[description = "Die count"] die_count: Option<u64>,
+    #[description = "Die size"] die_size: Option<u32>,
+    #[description = "Die count"] die_count: Option<u32>,
 ) -> Result<(), Error> {
     let die_size = die_size.unwrap_or(6);
     let die_count = die_count.unwrap_or(1);
@@ -21,8 +21,9 @@ pub async fn roll(
     let roll: u64 = {
         let mut rng = thread_rng();
         let uniform = rand::distributions::Uniform::new_inclusive(1, die_size);
-        (0..die_count).map(|_| uniform.sample(&mut rng)).sum()
+        (0..die_count).map(|_| u64::from(uniform.sample(&mut rng))).sum()
     };
+
 
     ctx.send(
         poise::CreateReply::default().embed(serenity::CreateEmbed::default().description(format!(
